perf: opt/llc -O3 + fix all C/C++ benchmarks to use int64_t
opt pipeline: -O2 → -O3 and llc -O=2 → -O=3. With O3, LLVM enables
more aggressive inlining, loop unrolling, and auto-vectorization.
Result: vāṇī now faster than C in 6/10 benchmarks.

benchmarks: all C/C++ files used bare `long` which is 32-bit on
Windows (MinGW), giving wrong outputs and an unfair 2× SIMD advantage
(32-bit fits 2× more elements per AVX2 register than 64-bit). Changed
all benchmark files to use `int64_t` / `PRId64` throughout:
  fib.{c,cpp}, sieve.{c,cpp}, matmul.{c,cpp}, sort.{c,cpp},
  graph_index.{c,cpp}, graph_weakptr.cpp, parsum.{c,cpp},
  hash.cpp, list.{c,cpp}, alloc.{c,cpp}, stats.{c,cpp}

After fix, fair 64-bit vs 64-bit comparison:
  vāṇī faster: matrix (0.97×), parsum (0.83×), graph (0.92×),
               linked list (0.86×), stats (0.96×), sort (0.56×)
  parity:      alloc (1.03×)
  small gap:   hashmap (1.12×), sieve (1.22×)
  real gap:    fibonacci (1.71×) — recursive call overhead

+++ benchmarks/06_parallel_sum/parsum.cpp
@@ -1,19 +1,21 @@
 // Benchmark 06 — Parallel sum of 50 000 000 elements  (C++ + OpenMP)
-// g++ -O2 -std=c++17 -fopenmp -o parsum_cpp parsum.cpp && ./parsum_cpp
+// g++ -O3 -march=native -std=c++17 -fopenmp -o parsum_cpp parsum.cpp && ./parsum_cpp
+#include <cinttypes>
+#include <cstdint>
 #include <cstdio>
 #include <vector>
 
 int main() {
-    const long N = 50000000L;
-    std::vector<long> data(N);
-    for (long i = 0; i < N; i++) data[i] = i % 1000;
+    const int64_t N = 50000000LL;
+    std::vector<int64_t> data((size_t)N);
+    for (int64_t i = 0; i < N; i++) data[(size_t)i] = i % 1000;
 
-    long total = 0;
+    int64_t total = 0;
 #ifdef _OPENMP
 #pragma omp parallel for reduction(+:total)
 #endif
-    for (long i = 0; i < N; i++)
-        total += data[i];
+    for (int64_t i = 0; i < N; i++)
+        total += data[(size_t)i];
 
-    std::printf("%ld\n", total);
+    std::printf("%" PRId64 "\n", total);
 }
